requirements.txt
amqp==5.0.6
anyio==3.3.0
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
Django==3.2.5
django-3-jet==1.0.8
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-rest-passwordreset==1.2.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.7.2
drf-friendly-errors-2==0.14.3
drf-reset-password==1.2.9
drf-yasg==1.20.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
parsel==1.6.0
Pillow==8.3.1
prompt-toolkit==3.0.19
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
